+++ CHANGELOG.md
@@ -17,10 +17,11 @@
         - Fixed various show/hide issues when several popups involved
         - Fixed popup closing when a click performs on/inside a trigger element while the popup is visible, now it stays visible
         - Fixed z-index when several popups are shown
+    - Added button views: `button`, `button-primary`, `button-danger` and `button-warning`
     - Added `select` view
     - Added `signature` view
-    - Improved `struct` view to show an object or an array signature when expanded and `[i]` action button is hovered
     - Added `limit` and `emptyText` options for `menu` view
+    - Improved `struct` view to show an object or an array signature when expanded and `[i]` action button is hovered
     - Improved `tree` view:
         - Added `limitLines` option to limit lines count on root children render, limits to 25 lines by default; use `limitLines: false` to disable it
         - Changed `limit` option to limit children count (this option served as `limitLines` before)
@@ -28,7 +29,7 @@
         - Added `collapsable` option to hide a toggle element when a leaf is expaned
         - Prevented large tree growing by keeping already visited data leafs collapsed
     - Added signature preset on report page
-    - Added button views: `button`, `button-primary`, `button-danger` and `button-warning`
+    - Added view editor input prettify button on report page
 
 ## 1.0.0-beta.8 (30-01-2019)